refactor: removeSectionByStation를 Sections로 이동함에 따른 테스트 코드 수정
