+++ src/test/java/nextstep/subway/section/domain/SectionTest.java
@@ -94,70 +94,4 @@ class SectionTest {
                 () -> 첫번째_구간.validateSectionAndAddSection(15, new Station(3L, "이매역"), new Station(4L, "수내역"), sections)
         );
     }
-
-    @Test
-    @DisplayName("시작 구간에 존재하는 역을 제거한다.")
-    void removeStartEdgeStation_test() {
-        //given
-        Station 야탑역 = new Station(3L, "야탑역");
-        Station 서현역 = new Station(4L, "서현역");
-
-        Section 두번째구간 = new Section(야탑역, 이매역, 15);
-
-        ArrayList<Section> sections = new ArrayList<>();
-        sections.add(첫번째_구간);
-        sections.add(두번째구간);
-        sections.add(new Section(이매역, 서현역, 5));
-
-        //when
-        첫번째_구간.removeSectionByStation(sections, 판교역, 0);
-
-        //then
-        assertThat(sections.get(0)).isEqualTo(두번째구간);
-    }
-
-    @Test
-    @DisplayName("종착역을 제거한다.")
-    void removeEndEdgeStation_test() {
-        //given
-        Station 야탑역 = new Station(3L, "야탑역");
-        Station 서현역 = new Station(4L, "서현역");
-
-        Section 두번째구간 = new Section(이매역, 야탑역, 15);
-        Section 세번째구간 = new Section(야탑역, 서현역, 5);
-
-        ArrayList<Section> sections = new ArrayList<>();
-        sections.add(첫번째_구간);
-        sections.add(두번째구간);
-        sections.add(세번째구간);
-
-        //when
-        첫번째_구간.removeSectionByStation(sections, 서현역, 2);
-
-        //then
-        assertThat(sections.get(1)).isEqualTo(두번째구간);
-    }
-
-    @Test
-    @DisplayName("중간역을 제거한다.")
-    void removeMiddleEdgeStation_test() {
-        //given
-        Station 야탑역 = new Station(3L, "야탑역");
-        Station 서현역 = new Station(4L, "서현역");
-
-        Section 두번째구간 = new Section(2L, 이매역, 야탑역, 15);
-        Section 세번째구간 = new Section(3L, 야탑역, 서현역, 5);
-
-        ArrayList<Section> sections = new ArrayList<>();
-        sections.add(첫번째_구간);
-        sections.add(두번째구간);
-        sections.add(세번째구간);
-
-        //when
-        두번째구간.removeSectionByStation(sections, 야탑역, 1);
-
-        //then
-        Section createdMiddleSection = new Section(2L, 판교역, 야탑역, 25);
-        assertThat(sections.get(0)).isEqualTo(createdMiddleSection);
-    }
 }
