+++ src/test/java/nextstep/subway/section/domain/SectionsTest.java
@@ -1,7 +1,9 @@
 package nextstep.subway.section.domain;
 
 import nextstep.subway.exception.CannotRemoveSingleSectionException;
+import nextstep.subway.line.domain.Line;
 import nextstep.subway.station.domain.Station;
+import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -71,6 +73,83 @@ class SectionsTest {
         //when
         assertThrows(CannotRemoveSingleSectionException.class,
                 () -> sections.validateAndRemoveSectionByStation(판교역)
+        );
+    }
+
+    @Test
+    @DisplayName("시작 구간에 존재하는 역을 제거한다.")
+    void removeStartEdgeStation_test() {
+        //given
+        Line 첫번째_라인 = new Line("1호선", "red");
+
+        Station 판교역 = new Station(1L, "판교역");
+        Station 이매역 = new Station(2L, "이매역");
+        Section 첫번째_구간 = new Section(1L, 판교역, 이매역, 첫번째_라인, 10);
+
+        Station 야탑역 = new Station(3L, "야탑역");
+        Section 두번째_구간 = new Section(2L, 야탑역, 이매역, 첫번째_라인, 15);
+
+        Sections sections = new Sections(Lists.newArrayList(첫번째_구간, 두번째_구간));
+
+        //when
+        sections.removeSectionByStation(판교역, 0);
+
+        //then
+        assertThat(sections.getSections()).containsExactly(두번째_구간);
+    }
+
+    @Test
+    @DisplayName("종착역을 제거한다.")
+    void removeEndEdgeStation_test() {
+        //given
+        Line 첫번째_라인 = new Line("1호선", "red");
+
+        Station 판교역 = new Station(1L, "판교역");
+        Station 야탑역 = new Station(3L, "야탑역");
+        Station 서현역 = new Station(4L, "서현역");
+        Station 이매역 = new Station(2L, "이매역");
+
+        Section 첫번째_구간 = new Section(1L, 판교역, 이매역, 첫번째_라인, 10);
+        Section 두번째_구간 = new Section(이매역, 야탑역, 15);
+        Section 세번째_구간 = new Section(야탑역, 서현역, 5);
+
+        Sections sections = new Sections(Lists.newArrayList(첫번째_구간, 두번째_구간, 세번째_구간));
+
+        //when
+        sections.removeSectionByStation(서현역, 2);
+
+        //then
+        assertThat(sections.getSections()).containsExactly(
+                첫번째_구간,
+                두번째_구간
+        );
+    }
+
+    @Test
+    @DisplayName("중간역을 제거한다.")
+    void removeMiddleEdgeStation_test() {
+        //given
+        Line 첫번째_라인 = new Line("1호선", "red");
+
+        Station 판교역 = new Station(1L, "판교역");
+        Station 야탑역 = new Station(3L, "야탑역");
+        Station 서현역 = new Station(4L, "서현역");
+        Station 이매역 = new Station(2L, "이매역");
+
+        Section 첫번째_구간 = new Section(1L, 판교역, 이매역, 첫번째_라인, 10);
+        Section 두번째_구간 = new Section(2L, 이매역, 야탑역, 15);
+        Section 세번째_구간 = new Section(3L, 야탑역, 서현역, 5);
+
+        Sections sections = new Sections(Lists.newArrayList(첫번째_구간, 두번째_구간, 세번째_구간));
+
+        //when
+        sections.removeSectionByStation(야탑역, 1);
+
+        //then
+        Section createdMiddleSection = new Section(2L, 판교역, 야탑역, 25);
+        assertThat(sections.getSections()).containsExactly(
+                createdMiddleSection,
+                세번째_구간
         );
     }
 }
